requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
arxiv==2.2.0
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
av==13.1.0
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-search==1.0.0b2
azure-search-documents==11.5.2
backoff==2.2.1
beautifulsoup4==4.13.3
blockbuster==1.5.24
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
cloudpickle==3.1.1
cloup==3.0.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
ctranslate2==4.6.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==4.4.2
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==8.0.4
emoji==1.6.3
et_xmlfile==2.0.0
exa-py==1.14.9
executing==2.2.0
fastembed==0.7.1
fastembed-gpu==0.7.1
faster-whisper==1.1.1
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.4
forbiddenfruit==0.1.4
fpdf==1.7.2
fpdf2==2.8.3
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
glcontext==3.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.99.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.28.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isosurfaces==0.1.2
jedi==0.19.2
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaleido==1.0.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-anthropic==0.3.15
langchain-community==0.3.26
langchain-core==0.3.66
langchain-deepseek==0.1.3
langchain-experimental==0.3.4
langchain-google-genai==2.1.5
langchain-google-vertexai==2.0.26
langchain-groq==0.3.2
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.24
langchain-qdrant==0.2.0
langchain-tavily==0.2.3
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.8
langgraph-api==0.2.56
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.3
langgraph-prebuilt==0.2.2
langgraph-runtime-inmem==0.3.0
langgraph-sdk==0.1.70
langsmith==0.3.45
linkup-sdk==0.2.6
logistro==1.1.0
loguru==0.7.3
lxml==5.4.0
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
Markdown==3.8.2
markdown-it-py==3.0.0
markdownify==1.1.0
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.9.4
mdurl==0.1.2
mmh3==5.1.0
moderngl==5.12.0
moderngl-window==3.1.1
moviepy==2.2.1
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.5.0
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
noise==1.2.2
numexpr==2.11.0
numpy==2.3.0
oauthlib==3.3.0
olefile==0.47
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
-e git+https://github.com/aditya-ladawa/cs101.git@3b77ff3633170f8ad56fc40b9e55a29117e1cdbc#egg=open_deep_research
openai==1.88.0
openpyxl==3.1.5
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pdfkit==1.0.0
pexpect==4.9.0
pillow==11.2.1
pilmoji==2.0.4
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
portalocker==2.10.1
primp==0.15.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.1
pyparsing==3.2.3
pypdf==5.6.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scipy==1.16.0
screeninfo==0.8.1
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
srt==3.5.3
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.47.0
structlog==25.4.0
svgelements==1.9.6
sympy==1.14.0
tavily-python==0.7.7
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
unstructured==0.17.2
unstructured-client==0.37.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.15
uvicorn==0.34.3
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2025.6.9
zstandard==0.23.0
